+++ hammer-api/src/main/java/cn/featherfly/hammer/tpl/annotation/Template.java
@@ -45,13 +45,4 @@ public @interface Template {
      * @return template type
      */
     TplType type() default TplType.AUTO;
-
-    /**
-     * value is template
-     *
-     * @return value is template
-     * @deprecated 会使用专门的注解处理非模板的情况，后续删除此属性
-     */
-    @Deprecated
-    boolean isTemplate() default true;
 }